requirements.txt
numpy
opencv-python
requests
azure-storage-blob
pillow